+++ src/helpers.ts
@@ -58,11 +58,12 @@ export class Helper {
     return new Promise<void>((resolve, reject) => {
       fs.readdir(directory, (err, files) => {
         if (err) reject(err);
-
-        for (const file of files) {
-          fs.unlink(path.join(directory, file), (err) => {
-            if (err) reject(err);
-          });
+        if (files) {
+          for (const file of files) {
+            fs.unlink(path.join(directory, file), (err) => {
+              if (err) reject(err);
+            });
+          }
         }
         resolve();
       });
